Rename testqueue to job_queue for coverage report
Coverage report did not include testqueue, so giving a better name for
that module. This way, coverage is giving proper results.
